requirements.txt
# Flask 웹 프레임워크
Flask==3.1.1
python-dotenv==1.1.1
gunicorn==23.0.0

# API 및 데이터 처리
requests==2.32.4
xmltodict==0.14.2

# LangChain AI 체인
langchain==0.3.27
langchain-openai==0.3.29
langchain-community==0.3.27

# 벡터 검색 (FAISS, BM25)
faiss-cpu==1.11.0.post1
rank-bm25==0.2.2

# OpenAI API
openai==1.99.9